# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AssigneeOut": {
        "properties": {
          "can_review": {
            "title": "Can Review",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "project_id": {
            "title": "Project Id",
            "type": "integer"
          },
          "role": {
            "enum": [
              "owner",
              "admin",
              "member"
            ],
            "title": "Role",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "user_id",
          "username",
          "role",
          "can_review",
          "status",
          "created_at"
        ],
        "title": "AssigneeOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewCreate": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "result": {
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Result",
            "type": "string"
          }
        },
        "required": [
          "result"
        ],
        "title": "ReviewCreate",
        "type": "object"
      },
      "ReviewOut": {
        "properties": {
          "annotation_revision": {
            "title": "Annotation Revision",
            "type": "integer"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detection_import_revision": {
            "default": 0,
            "title": "Detection Import Revision",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "identity_revision": {
            "default": 0,
            "title": "Identity Revision",
            "type": "integer"
          },
          "project_id": {
            "title": "Project Id",
            "type": "integer"
          },
          "result": {
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Result",
            "type": "string"
          },
          "reviewer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "submission_annotations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SubmissionAnnotationSnapshotOut"
            },
            "title": "Submission Annotations",
            "type": "array"
          },
          "submission_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Id"
          },
          "video_id": {
            "title": "Video Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "project_id",
          "video_id",
          "result",
          "annotation_revision",
          "created_at"
        ],
        "title": "ReviewOut",
        "type": "object"
      },
...
      "RoleDefinitionOut": {
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "max_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Count"
          },
          "min_count": {
            "title": "Min Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role_sort_order": {
            "title": "Role Sort Order",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "name",
          "min_count",
          "role_sort_order"
        ],
        "title": "RoleDefinitionOut",
        "type": "object"
      },
...
      "SubmissionAnnotationSnapshotOut": {
        "properties": {
          "category_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Group"
          },
          "category_id": {
            "title": "Category Id",
            "type": "integer"
          },
          "category_name": {
            "title": "Category Name",
            "type": "string"
          },
          "category_participant_mode": {
            "default": "unordered",
            "enum": [
              "unordered",
              "role_based"
            ],
            "title": "Category Participant Mode",
            "type": "string"
          },
          "end_frame": {
            "title": "End Frame",
            "type": "integer"
          },
          "end_time": {
            "title": "End Time",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "mouse_ids": {
            "default": [],
            "items": {
              "type": "integer"
            },
            "title": "Mouse Ids",
            "type": "array"
          },
          "participant_roles": {
            "additionalProperties": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "default": {},
            "title": "Participant Roles",
            "type": "object"
          },
          "role_definitions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RoleDefinitionOut"
            },
            "title": "Role Definitions",
            "type": "array"
          },
          "source_annotation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Annotation Id"
          },
          "start_frame": {
            "title": "Start Frame",
            "type": "integer"
          },
          "start_time": {
            "title": "Start Time",
            "type": "number"
          }
        },
        "required": [
          "id",
          "category_id",
          "category_name",
          "start_time",
          "end_time",
          "start_frame",
          "end_frame"
        ],
        "title": "SubmissionAnnotationSnapshotOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VideoOut": {
        "properties": {
          "annotation_revision": {
            "default": 1,
            "title": "Annotation Revision",
            "type": "integer"
          },
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "assignee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssigneeOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "assignee_membership_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Membership Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detection_import_revision": {
            "default": 0,
            "title": "Detection Import Revision",
            "type": "integer"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "fps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fps"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "identity_revision": {
            "default": 0,
            "title": "Identity Revision",
            "type": "integer"
          },
          "playback_status": {
            "enum": [
              "ready",
              "pending",
              "failed",
              "unavailable"
            ],
            "title": "Playback Status",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submission_annotations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SubmissionAnnotationSnapshotOut"
            },
            "title": "Submission Annotations",
            "type": "array"
          },
          "submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "workflow_status": {
            "default": "draft",
            "title": "Workflow Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "filename",
          "playback_status",
          "status",
          "created_at"
        ],
        "title": "VideoOut",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Behavior Annotation Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/{project_id}/reviews/queue": {
      "get": {
        "operationId": "review_queue_api_projects__project_id__reviews_queue_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VideoOut"
                  },
                  "title": "Response Review Queue Api Projects  Project Id  Reviews Queue Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Queue",
        "tags": [
          "reviews"
        ]
      }
    },
...
    "/api/projects/{project_id}/videos/{video_id}/review": {
      "post": {
        "operationId": "create_review_api_projects__project_id__videos__video_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Review",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/projects/{project_id}/videos/{video_id}/reviews": {
      "get": {
        "operationId": "review_history_api_projects__project_id__videos__video_id__reviews_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewOut"
                  },
                  "title": "Response Review History Api Projects  Project Id  Videos  Video Id  Reviews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review History",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/projects/{project_id}/videos/{video_id}/submit": {
      "post": {
        "operationId": "submit_video_api_projects__project_id__videos__video_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Video",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/projects/{project_id}/videos/{video_id}/withdraw": {
      "post": {
        "operationId": "withdraw_video_api_projects__project_id__videos__video_id__withdraw_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Withdraw Video",
        "tags": [
          "reviews"
        ]
      }
    },
...
  }
}